Extract comparison methods to ComparisonService
